[AMD] Update nightly test configurations and add new accuracy tests
- Modified the nightly test workflow to include new 1-GPU and 2-GPU accuracy tests for MI30x models.
- Removed deprecated 2-GPU tests and adjusted job filters for MI35x architecture.
- Introduced new test files for GSM8K evaluation on both 1-GPU and 2-GPU configurations.
- Enhanced CI registration for accuracy tests to ensure comprehensive coverage across AMD architectures.
